Add null check for `client`
The `client` can be null occasionally when windowActivated is triggered (not exactly sure why) and producing the below warning message:

file:///usr/share/kwin/scripts/kwin-system76-scheduler-integration/contents/code/main.js:5: TypeError: Cannot read property 'pid' of null

While this warning is not harmful, it is polluting the logs. The warning should not appearing anymore with the null check.